Refactor code to package
